requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
Faker==24.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
regex==2024.4.28
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
validate-docbr==1.10.0
wheel==0.43.0
